Fix warnings

+++ src/VDP.hpp
@@ -109,16 +109,16 @@ public:
         switch (state) {
             case VDPstate::VRAMwrite: { // TODO: word write
                 if (sizeof(DataType) == 4) {
-                    vram[currentAddr] = static_cast<uint8>(data & 0xFF);
-                    vram[currentAddr + 1] = static_cast<uint8>((data >> 8) & 0xFF);
-                    vram[currentAddr + 2] = static_cast<uint8>((data >> 16) & 0xFF);
-                    vram[currentAddr + 3] = static_cast<uint8>(data >> 24);
+                    vram[currentAddr] = static_cast<uint8>(static_cast<uint32>(data) & 0xFF);
+                    vram[currentAddr + 1] = static_cast<uint8>((static_cast<uint32>(data) >> 8) & 0xFF);
+                    vram[currentAddr + 2] = static_cast<uint8>((static_cast<uint32>(data) >> 16) & 0xFF);
+                    vram[currentAddr + 3] = static_cast<uint8>(static_cast<uint32>(data) >> 24);
 
                     currentAddr += 4; // autoincrement
                 }
                 else if (sizeof(DataType) == 2) {
-                    vram[currentAddr] = static_cast<uint8>(data & 0xFF);
-                    vram[currentAddr + 1] = static_cast<uint8>((data >> 8) & 0xFF);
+                    vram[currentAddr] = static_cast<uint8>(static_cast<uint16>(data) & 0xFF);
+                    vram[currentAddr + 1] = static_cast<uint8>((static_cast<uint16>(data) >> 8) & 0xFF);
 
                     currentAddr += 2; // autoincrement
                 }
